Add a lifecycle consumer to ensure ordering
Before sending the `render` event to the consumer, make sure that the
consumer has received the call to `activate` and has had the opportunity
to set up any state that requires its dependencies.

This is implemented via a new lifecycle service that expects to receive
an activate event from the child sandbox once it is ready to receive
other events.