Added ad card.
The ad card shows an ad in the ‘thumbnail’ intent, and when it's done
transitions into a ‘takingSurvey’  intent.  It propagates the ‘surveyTaken’
intent along with the video id.

The ad card cheats at the moment, as it knows the aspect ratio of the video
thumbnails.  This would not be necessary if we had ALLOWSEAMLESS or a fallback
height service (e.g. based on DOM Mutation observers).

Wiretaps don't propagate: the playground needs a simple wiretap propagation
service.