Add support for IE10, Firefox.
BREAKING CHANGES:
  - `card.then` → `card.promise.then`
  - `conductor.load` does not start the sandbox.  The sandbox is started during
    `appendTo`
  - `initializeDOM` is now required to call `appendTo` for each child card, for
    the above reason.  The default `initializeDOM` appends all children to
    `document.body`, in order.
  - `oasisURL` must now refer to the polyglot file `conductor.js-0.1.0.js.html`.
    This file does not need to end in `html` but it does need to be served with
    an html content type.

Updates Oasis and RSVP.

This commit should also add support for IE9, although the Superbowl demo behaves
slightly differently between IE10 and IE9.  In the latter, the video does not
transition to the survey once it finishes playing: this might be a problem with
the YouTube API however.

There remains some CSS differences between IE and Chrome in the playground.